refactor(app-deployment) :: ResourceInfo 메모리 타입을 Double로 변경

+++ src/main/java/madp/appdeployment/domain/domain/vo/ResourceInfo.java
@@ -16,13 +16,13 @@ public class ResourceInfo {
     private Double cpu;
 
     @Column(name = "resource_memory", nullable = false)
-    private Integer memory;
+    private Double memory;
 
     @Column(name = "resource_disk", nullable = false)
     private Integer disk;
 
     @Builder
-    public ResourceInfo(Double cpu, Integer memory, Integer disk) {
+    public ResourceInfo(Double cpu, Double memory, Integer disk) {
         validateFields(cpu, memory, disk);
 
         this.cpu = cpu;
@@ -31,7 +31,7 @@ public class ResourceInfo {
     }
 
     // ConfigurationProperties로 정책에 따라서 validate 함수 바꿀 예정
-    private void validateFields(Double cpu, Integer memory, Integer disk) {
+    private void validateFields(Double cpu, Double memory, Integer disk) {
         if (cpu == null) {
             throw new InvalidResourceInfoException("CPU는 필수입니다.");
         }
